Update README.md
This program contains distinct functions forsorting a list of integers. The program uses a command line interface. 
The program recieves a list of integers,sorts the list using a method specified by the user and prints the list to the screen.

Made by WGDEV, some rights reserved, see license.txt for more info